+++ include/miopen/miopen.h
@@ -8256,7 +8256,7 @@ MIOPEN_EXPORT miopenStatus_t miopenPadConstantFwd(miopenHandle_t handle,
                                                   miopenTensorDescriptor_t yDesc,
                                                   const void* x,
                                                   void* y,
-                                                  const size_t* padding,
+                                                  const int64_t* padding,
                                                   const int padding_size,
                                                   float value);
 
@@ -8281,7 +8281,7 @@ MIOPEN_EXPORT miopenStatus_t miopenPadConstantBwd(miopenHandle_t handle,
                                                   miopenTensorDescriptor_t dyDesc,
                                                   void* dx,
                                                   const void* dy,
-                                                  const size_t* padding,
+                                                  const int64_t* padding,
                                                   const int padding_size);
 
 /** @} */
